+++ include/markov_processor.hpp
@@ -9,8 +9,8 @@ class MarkovProcessor {
 public:
   MarkovProcessor(MarkovChain &mc, const std::filesystem::path &build_folder, const std::filesystem::path &output_path,
                   const std::filesystem::path &latex_output_directory, const std::filesystem::path &filelist_path,
-                  const std::string &latex_compiler, const std::string &latex_compiler_options, bool edit_latex,
-                  bool verbose, bool no_cleanup);
+                  const std::string &file_extension, const std::string &latex_compiler,
+                  const std::string &latex_compiler_options, bool edit_latex, bool verbose, bool no_cleanup);
 
   void video_and_build(const std::filesystem::path &video_folder, std::size_t iterations);
 
@@ -26,6 +26,7 @@ private:
   const std::filesystem::path &output_path;
   const std::filesystem::path &latex_output_directory;
   const std::filesystem::path &filelist_path;
+  const std::string &file_extension;
   const std::string &latex_compiler;
   const std::string &latex_compiler_options;
   bool edit_latex;