+++ src/markov_processor.cpp
@@ -11,11 +11,13 @@ namespace fs = std::filesystem;
 
 MarkovProcessor::MarkovProcessor(MarkovChain &mc, const fs::path &build_folder, const fs::path &output_path,
                                  const fs::path &latex_output_directory, const fs::path &filelist_path,
-                                 const std::string &latex_compiler, const std::string &latex_compiler_options,
-                                 bool edit_latex, bool verbose, bool no_cleanup)
+                                 const std::string &file_extension, const std::string &latex_compiler,
+                                 const std::string &latex_compiler_options, bool edit_latex, bool verbose,
+                                 bool no_cleanup)
     : mc(mc), build_folder(build_folder), output_path(output_path), latex_output_directory(latex_output_directory),
-      filelist_path(filelist_path), latex_compiler(latex_compiler), latex_compiler_options(latex_compiler_options),
-      edit_latex(edit_latex), verbose(verbose), no_cleanup(no_cleanup) {}
+      filelist_path(filelist_path), file_extension(file_extension), latex_compiler(latex_compiler),
+      latex_compiler_options(latex_compiler_options), edit_latex(edit_latex), verbose(verbose), no_cleanup(no_cleanup) {
+}
 
 void MarkovProcessor::video_and_build(const fs::path &video_folder, std::size_t iterations) {
   const auto &markov_states = iterate_markov_states(mc, iterations);
@@ -31,8 +33,8 @@ void MarkovProcessor::video_and_build(const fs::path &video_folder, std::size_t 
   compile_all_markov_graphs(build_folder, transition_matrix_size, latex_output_directory, latex_compiler,
                             latex_compiler_options, verbose);
   convert_all_pdfs_to_pngs(build_folder / latex_output_directory, transition_matrix_size, build_folder, verbose);
-  overlay_images_to_videos(video_folder, build_folder, transition_matrix_size, build_folder, verbose);
-  create_filelist(markov_states, build_folder / filelist_path);
+  overlay_images_to_videos(video_folder, file_extension, build_folder, transition_matrix_size, build_folder, verbose);
+  create_filelist(markov_states, build_folder / filelist_path, file_extension);
   combine_segments(build_folder / filelist_path, output_path, verbose);
 
   if (!no_cleanup)
@@ -52,8 +54,8 @@ void MarkovProcessor::video_only(const fs::path &video_folder, std::size_t itera
   compile_all_markov_graphs(build_folder, transition_matrix_size, latex_output_directory, latex_compiler,
                             latex_compiler_options, verbose);
   convert_all_pdfs_to_pngs(build_folder / latex_output_directory, transition_matrix_size, build_folder, verbose);
-  overlay_images_to_videos(video_folder, build_folder, transition_matrix_size, build_folder, verbose);
-  create_filelist(markov_states, build_folder / filelist_path);
+  overlay_images_to_videos(video_folder, file_extension, build_folder, transition_matrix_size, build_folder, verbose);
+  create_filelist(markov_states, build_folder / filelist_path, file_extension);
   combine_segments(build_folder / filelist_path, output_path, verbose);
 
   if (!no_cleanup)